requirements.txt
coverage==3.7.1
httplib2==0.8
pytest==2.5.2
pytest-xdist==1.11
wsgiref==0.1.2
futures==2.1.3
SQLAlchemy==1.0.4
pymongo==2.5.2  # For json_util in bson
dnspython==1.11.1
boto==2.10.0
ipython==1.0.0
Flask==0.10.1
gevent==1.0.1
gevent-socketio==0.3.5-rc2
geventconnpool==0.2.1
gunicorn==19.0.0
colorlog==1.8
MySQL-python==1.2.5
requests==2.4.3
chardet==2.1.1
setproctitle==1.1.8
gdata==2.0.18
git+https://github.com/nylas/dateutil.git#egg=python-dateutil
flanker==0.4.28
git+https://github.com/nylas/imapclient.git@master#egg=imapclient
alembic==0.6.4
iconv==1.0
pep8==1.4.6
flake8==2.1.0
pyflakes==0.7.3
pylint==1.1.0
html2text==2014.9.8
python-magic==0.4.6
greenlet==0.4.2
PyMySQL==0.6.2
click==4.0
faulthandler==2.3
pyinstrument==0.12
structlog==0.4.1
raven==5.0.0
simplejson==3.6.0
Flask-RESTful==0.3.2
uritemplate==0.6
icalendar==3.8.2
git+https://github.com/kav-ya/pynacl.git@libsodium_fix#egg=pynacl
pyaml==14.05.7
enum34==1.0.4
msgpack-python==0.4.2
guppy==0.1.10
elasticsearch==1.2.0
redis==2.10.3
hiredis==0.1.5
mock==1.0.1
mockredispy==2.9.0.10
vobject==0.8.1c
lxml==3.4.2
arrow==0.5.4
cryptography==0.8.2
Twisted==15.1.0
hypothesis==1.4.1
service_identity
